requirements.txt
asgiref==3.8.1
Django>=5.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.14,<4.0
django-elasticsearch-dsl>=8.0.0
elasticsearch>=8.13.0,<9.0.0
django-extensions